Renamed almost all files to our conventions.
