+++ gettsim/tests/test_zu_versteuerndes_eink.py
@@ -7,7 +7,7 @@ from pandas.testing import assert_series_equal
 
 from gettsim.config import ROOT_DIR
 from gettsim.pre_processing.policy_for_date import get_policies_for_date
-from gettsim.taxes.zve import zve
+from gettsim.taxes.zu_versteuerndes_eink import zve
 
 
 IN_COLS = [
